llutils: Incorporated Nick's suggestion on avoiding try-except for removing the lock file

+++ pylal/pylal/pylal_exttrig_llutils.py
@@ -287,10 +287,8 @@ def del_lock():
    """
    Removes the lock file
    """
-   try:
+   if os.path.exists(get_lockname()):
      os.remove(get_lockname())
-   except:
-     pass
    info('monitor','Program exit normally')   
 
 # --------------------------------------
